renames 'utils' dir to 'modules'
