+++ assets/vue/modules/colors.ts
