+++ assets/vue/modules/constants.ts
