+++ assets/vue/modules/index.ts
